# gamemarket/src/main/resources/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/v1/products/{id}": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailsDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "product-controller"
        ],
        "operationId": "updateProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDetailsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailsDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "product-controller"
        ],
        "operationId": "deleteProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "operationId": "getProducts",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailsListDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "product-controller"
        ],
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDetailsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{customerReference}/activate": {
      "post": {
        "tags": [
          "product-controller"
        ],
        "operationId": "getProductByKeyActivation",
        "parameters": [
          {
            "name": "customerReference",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers": {
      "get": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "getAllCustomers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDetailsListDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "createCustomer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerDetailsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDetailsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/{id}": {
      "get": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "getCustomerById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDetailsDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProductDetailsDto": {
        "required": [
          "developer",
          "deviceTypes",
          "genres",
          "shortDescription",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 98,
            "minLength": 0,
            "type": "string"
          },
          "shortDescription": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "price": {
            "minimum": 1,
            "type": "number",
            "format": "double"
          },
          "developer": {
            "maxLength": 101,
            "minLength": 0,
            "type": "string"
          },
          "deviceTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CustomerDetailsDto": {
        "required": [
          "email",
          "name",
          "phoneNumber",
          "region"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 99,
            "minLength": 0,
            "type": "string"
          },
          "phoneNumber": {
            "pattern": "^\\+?[0-9]{10,15}$",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "region": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "deviceTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ProductDetailsEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "developer": {
            "type": "string"
          },
          "deviceTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ProductDetailsListDto": {
        "type": "object",
        "properties": {
          "productDetailsEntries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDetailsEntry"
            }
          }
        }
      },
      "CustomerDetailsEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "deviceTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CustomerDetailsListDto": {
        "type": "object",
        "properties": {
          "customerDetailsEntries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerDetailsEntry"
            }
          }
        }
      }
    }
  }
}